+++ lib/i18n/translations_es.g.dart
@@ -257,6 +257,7 @@ class _TranslationsScriptEditorPageEs extends TranslationsScriptEditorPageEn {
 	@override String get insert_variable_title => 'Insertar una predefinida variable';
 	@override String get choice_common_constants => 'Constantes comunes';
 	@override String get choice_script_variables => 'Variable para el tipo de código';
+	@override String get invalid_script => 'El código tiene errores. Por favor, corregirlos.';
 }
 
 // Path: variables_table