+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * Rename window title in desktop version
 * Set board reversed by default if player has black side
 
+* Showing several errors at once instead of just the first error for scripts
+* Caution ! The scripts checker does not warn you any more for being too restrictive
 * (Android) (bug fix) : sometimes could not run long scripts
 * (bug fix) : script types in some errors were not precise enough : sometimes we missed the piece kind
 * (bug fix) : script error "parenthesis without expression" made program crash instead of reporting it